+++ db-queue-scheduler-core/src/main/java/ru/yoomoney/tech/dbqueue/scheduler/config/impl/LoggingScheduledTaskLifecycleListener.java
@@ -3,6 +3,7 @@ package ru.yoomoney.tech.dbqueue.scheduler.config.impl;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import ru.yoomoney.tech.dbqueue.scheduler.config.ScheduledTaskLifecycleListener;
+import ru.yoomoney.tech.dbqueue.scheduler.models.ScheduledTaskContext;
 import ru.yoomoney.tech.dbqueue.scheduler.models.ScheduledTaskExecutionResult;
 import ru.yoomoney.tech.dbqueue.scheduler.models.ScheduledTaskIdentity;
 
@@ -20,12 +21,13 @@ public class LoggingScheduledTaskLifecycleListener implements ScheduledTaskLifec
     private static final Logger log = LoggerFactory.getLogger(LoggingScheduledTaskLifecycleListener.class);
 
     @Override
-    public void started(@Nonnull ScheduledTaskIdentity taskIdentity) {
+    public void started(@Nonnull ScheduledTaskIdentity taskIdentity, @Nonnull ScheduledTaskContext taskContext) {
         log.info("task started: identity={}", taskIdentity.asString());
     }
 
     @Override
     public void finished(@Nonnull ScheduledTaskIdentity taskIdentity,
+                         @Nonnull ScheduledTaskContext taskContext,
                          @Nonnull ScheduledTaskExecutionResult executionResult,
                          @Nonnull Instant nextExecutionTime,
                          long processTaskTimeInMills) {
@@ -39,7 +41,9 @@ public class LoggingScheduledTaskLifecycleListener implements ScheduledTaskLifec
     }
 
     @Override
-    public void crashed(@Nonnull ScheduledTaskIdentity taskIdentity, @Nullable Throwable exc) {
+    public void crashed(@Nonnull ScheduledTaskIdentity taskIdentity,
+                        @Nonnull ScheduledTaskContext taskContext,
+                        @Nullable Throwable exc) {
         log.error("task crashed: identity={}", taskIdentity.asString(), exc);
     }
 }